requirements.txt
opencv_contrib_python==4.9.0.80
opencv_python_headless==4.9.0.80
openpyxl==3.1.2
pandas==2.2.2
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.10
streamlit==1.33.0
